% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_choose_plants.R
\name{compose_filters}
\alias{compose_filters}
\title{Compose filters}
\usage{
compose_filters(select_filters, checkbox_filters, zone)
}
\arguments{
\item{select_filters}{pickerInput filters}

\item{checkbox_filters}{checkboxInput filters}

\item{zone}{Hardiness zone}
}
\value{
A list with
\item{url_args}{URL arguments that can be passed to a species list request}
\item{addl_filters}{Additional filters to be applied after the species list
request has been performed}
}
\description{
Combine pickerInput filters and checkbox filters
}
