% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_choose_plants.R
\name{insert_plant_cards}
\alias{insert_plant_cards}
\title{Insert plant cards}
\usage{
insert_plant_cards(ns, details, border_ids)
}
\arguments{
\item{ns}{Namespacing function}

\item{details}{Named list of species details}

\item{border_ids}{IDs of the plants in your border}
}
\description{
Add plant cards to the UI
}
