% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_choose_plants.R
\name{make_filters}
\alias{make_filters}
\title{Make filters}
\usage{
make_filters(id, ns)
}
\arguments{
\item{id}{Filter ID}

\item{ns}{Namespacing function}
}
\value{
The UI for a pickerInput filter
}
\description{
Make UI for a pickerInput filter
}
