% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_border_info.R
\name{make_guide_content}
\alias{make_guide_content}
\title{Make guide content}
\usage{
make_guide_content(ns, care_guide, details_record, side)
}
\arguments{
\item{ns}{Namespacing function}

\item{care_guide}{The care guide data for the plant}

\item{details_record}{The details record for the plant}

\item{side}{The side of the page the text should appear on relative to the
image}
}
\value{
The html for an entry in the plant guide
}
\description{
Make content for the plant care guide tab
}
