% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_choose_plants.R
\name{make_plant_stub}
\alias{make_plant_stub}
\title{Make plant stub}
\usage{
make_plant_stub(
  session,
  id,
  details_record,
  your_border,
  card_ids,
  search_ids,
  ns = NULL
)
}
\arguments{
\item{session}{Session object}

\item{details_record}{Species details for the plant stub}

\item{your_border}{A reactiveVal with plant details for your border}

\item{card_ids}{A reactiveVal with the IDs of plants currently shown in the
plant list}

\item{search_ids}{A reactiveVal with the IDs of plants currently shown in the
search results}

\item{ns}{Namespacing function, only used for stubs in the search results,
which live in the choose plants module}
}
\description{
Make a plant stub for the search results or border Offcanvas
}
