% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_req_perenual.R
\name{process_page_resp}
\alias{process_page_resp}
\title{Process page response}
\usage{
process_page_resp(page_resp, addl_filters)
}
\arguments{
\item{page_resp}{Page response from the species list endpoint}
}
\value{
List of
\item{data}{Any remaining data after filters are applied}
\item{page_count}{The page count. Only used for the first page.}
\item{status}{API status ("OK" or "api_error")}
}
\description{
Apply post hoc filters to a page response from the species list endpoint
}
