% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_req_perenual.R
\name{req_details}
\alias{req_details}
\title{Details request}
\usage{
req_details(id)
}
\arguments{
\item{id}{ID to request}
}
\value{
Response
}
\description{
Request a single plant from the details endpoint. Errors are suppressed
}
