% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_req_perenual.R
\name{req_details_batch}
\alias{req_details_batch}
\title{Request batch of details}
\usage{
req_details_batch(ids)
}
\arguments{
\item{ids}{IDs to request}
}
\value{
List of
\item{data}{A list of details data for successful requests.}
\item{rem_ids}{Remaining IDs that were not used.}
\item{status}{API status ("OK" or "api_error")}
}
\description{
Request a batch of details from the details endpoint.
}
