% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_req_perenual.R
\name{req_more_plants}
\alias{req_more_plants}
\title{Request more plants}
\usage{
req_more_plants(url_args, addl_filters, rem_pages)
}
\arguments{
\item{url_args}{Url arguments to include in request}

\item{addl_filters}{Additional filters to apply after the request}

\item{rem_pages}{Set of remaining pages that match the filters}
}
\value{
List of
\item{details}{A list of details data for successful requests.}
\item{rem_pages}{Remaining species list pages that were not used.}
\item{rem_ids}{Remaining IDs that were not used.}
\item{status}{API status ("OK" or "api_error")}
}
\description{
Higher level function that uses req_page_batch and req_details batch to try
to get another set of details.
}
