% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_req_perenual.R
\name{req_page_batch}
\alias{req_page_batch}
\title{Request batch of pages from species list}
\usage{
req_page_batch(url_args, addl_filters, rem_pages, pages_to_add = 3)
}
\arguments{
\item{url_args}{Url arguments to include in request}

\item{addl_filters}{Additional filters to apply after the request}

\item{rem_pages}{Set of remaining pages that match the filters}

\item{pages_to_add}{Number of pages to request. Defaults to three.}
}
\value{
List of
\item{data}{Combined species list data matching filters}
\item{ok_ids}{IDs of matching plants.}
\item{rem_pages}{Remaining pages that match the URL filters.}
\item{status}{API status ("OK" or "api_error")}
}
\description{
Request a batch of pages from species list using url filters. Then apply
post hoc filters.
}
