% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_req_perenual.R
\name{req_species_list}
\alias{req_species_list}
\title{Species list request}
\usage{
req_species_list(...)
}
\arguments{
\item{...}{URL filters accepted by the species list endpoint}
}
\value{
Request object
}
\description{
Request a single page from the species list
}
