% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_req_perenual.R
\name{resp_to_list}
\alias{resp_to_list}
\title{Convert details response to list}
\usage{
resp_to_list(resp)
}
\arguments{
\item{resp}{Response from the details endpoint}
}
\value{
List of plant details for one plant
}
\description{
Convert a response from the details endpoint to a list
}
