% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_choose_plants.R
\name{update_border}
\alias{update_border}
\title{Update border}
\usage{
update_border(session, id, your_border, details_record, card_ids, search_ids)
}
\arguments{
\item{session}{Session object}

\item{your_border}{A reactiveVal with plant details for your border}

\item{details_record}{Species details for the plant stub}

\item{card_ids}{A reactiveVal with the IDs of plants currently shown in the
plant list}

\item{search_ids}{A reactiveVal with the IDs of plants currently shown in the
search results}
}
\description{
Update border when a plant is added or removed. Takes care of updating any
relevant UI in the border offCanvas, search results, and plant card list.
}
