% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_choose_plants.R
\name{update_plant_list}
\alias{update_plant_list}
\title{Update plant list}
\usage{
update_plant_list(ns, details, your_border, card_ids, search_ids)
}
\arguments{
\item{ns}{Namespacing function}

\item{details}{Named list of species details}

\item{your_border}{A reactiveVal with plant details for your border}

\item{card_ids}{A reactiveVal with the IDs of plants currently shown in the
plant list}

\item{search_ids}{A reactiveVal with the IDs of plants currently shown in the
search results}
}
\description{
Update both the UI and the server logic for plant cards added to the plant
list
}
