% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_choose_plants.R
\name{update_show_more}
\alias{update_show_more}
\title{Update show more}
\usage{
update_show_more(rem_pages, rem_ids, status, any_results = FALSE)
}
\arguments{
\item{rem_pages}{Remaining pages from the species list request}

\item{rem_ids}{Remaining IDs from the last batch of species list pages}

\item{status}{API status ("OK" or "api_error")}

\item{any_result}{Boolean - are there any results already shown in the plant
list?}
}
\description{
Update the show more button
}
