% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_plant_entry}
\alias{valid_plant_entry}
\title{Validate plant entry}
\usage{
valid_plant_entry(details_record)
}
\arguments{
\item{details_record}{Details record for a plant}
}
\value{
TRUE if plant entry is valid, false if not
}
\description{
Check if a plant entry is valid on the fly to avoid errors
}
